VER3-232 refactor study service into study form
