+++ src/app/study-form/solve-for/solve-for.component.spec.ts
@@ -2,7 +2,7 @@ import {async, ComponentFixture, TestBed} from '@angular/core/testing';
 
 import { SolveForComponent } from './solve-for.component';
 import {ReactiveFormsModule} from '@angular/forms';
-import {StudyService} from '../../shared/study.service';
+import {StudyService} from '../study.service';
 import {Http} from '@angular/http';
 import { MockBackend } from '@angular/http/testing';
 import {DebugElement} from '@angular/core';
