+++ src/app/study-form/statistical-tests/statistical-tests.component.spec.ts
@@ -3,7 +3,7 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { StatisticalTestsComponent } from './statistical-tests.component';
 import {By} from '@angular/platform-browser';
 import {DebugElement} from '@angular/core';
-import {StudyService} from '../../shared/study.service';
+import {StudyService} from '../study.service';
 import {Http} from '@angular/http';
 import {MockBackend} from '@angular/http/testing';
