+++ src/app/study-form/target-event/target-event.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import {StudyService} from '../../shared/study.service';
+import {StudyService} from '../study.service';
 import {constants} from '../../shared/constants';
 import {Subscription} from 'rxjs/Subscription';
