+++ src/app/study-form/type-one-error/type-one-error.component.spec.ts
@@ -5,7 +5,7 @@ import { TypeOneErrorComponent } from './type-one-error.component';
 import {LoggerModule, NGXLogger, NGXLoggerMock} from 'ngx-logger';
 import {MockBackend} from '@angular/http/testing';
 import {Http} from '@angular/http';
-import {StudyService} from '../../shared/study.service';
+import {StudyService} from '../study.service';
 
 describe('TypeOneErrorComponent', () => {
   let component: TypeOneErrorComponent;
