+++ src/app/study-form/within-isu-outcomes/within-isu-outcomes.component.spec.ts
@@ -5,7 +5,7 @@ import {ReactiveFormsModule} from '@angular/forms';
 import {DebugElement} from '@angular/core';
 import {By} from '@angular/platform-browser';
 import {NavigationService} from '../../shared/navigation.service';
-import {StudyService} from '../../shared/study.service';
+import {StudyService} from '../study.service';
 import {MockBackend} from '@angular/http/testing';
 import {Http} from '@angular/http';
 import {Outcome} from "../../shared/Outcome";
